feat: fpe perform

+++ test/ecb_perform.c
@@ -11,7 +11,7 @@ static const unsigned char msg[16] = {0x01, 0x23, 0x45, 0x67, 0x89, 0xab, 0xcd, 
                                       0xfe, 0xdc, 0xba, 0x98, 0x76, 0x54, 0x32, 0x10};
 
 void test_aes_standard_perf() {
-    int i;
+    long long i;
     unsigned char cipher[16] = {0};
     clock_t program_start, program_end;
     double ts;
@@ -20,7 +20,7 @@ void test_aes_standard_perf() {
     WBCRYPTO_aes_init_key(aes_ctx, key, sizeof(key));
 
     program_start = clock();
-    for (i = 0; i < TEST_CYCLE_NUM * 64 * 1024; i++) {
+    for (i = 0; i < (long long)TEST_CYCLE_NUM * 64 * 1024; i++) {
         WBCRYPTO_aes_encrypt(msg, cipher, aes_ctx);
     }
     program_end = clock();
@@ -33,7 +33,7 @@ void test_aes_standard_perf() {
 }
 
 void test_wbaes_standard_perf() {
-    int i;
+    long long i;
     unsigned char cipher[16] = {0};
     clock_t program_start, program_end;
     double ts;
@@ -43,7 +43,7 @@ void test_wbaes_standard_perf() {
     WBCRYPTO_wbaes_gen_table(wbaes_ctx_enc, key, sizeof(key));
 
     program_start = clock();
-    for (i = 0; i < TEST_CYCLE_NUM * 64 * 1024; i++) {
+    for (i = 0; i < (long long)TEST_CYCLE_NUM * 64 * 1024; i++) {
         WBCRYPTO_wbaes_encrypt(msg, cipher, wbaes_ctx_enc);
     }
     program_end = clock();
@@ -56,7 +56,7 @@ void test_wbaes_standard_perf() {
 }
 
 void test_sm4_standard_perf() {
-    int i;
+    long long i;
     unsigned char cipher[16] = {0};
     clock_t program_start, program_end;
     double ts;
@@ -65,7 +65,7 @@ void test_sm4_standard_perf() {
     WBCRYPTO_sm4_init_key(sm4_ctx, key, sizeof(key));
 
     program_start = clock();
-    for (i = 0; i < TEST_CYCLE_NUM * 64 * 1024; i++) {
+    for (i = 0; i < (long long)TEST_CYCLE_NUM * 64 * 1024; i++) {
         WBCRYPTO_sm4_encrypt(msg, cipher, sm4_ctx);
     }
     program_end = clock();
@@ -78,7 +78,7 @@ void test_sm4_standard_perf() {
 }
 
 void test_wbsm4_standard_perf() {
-    int i;
+    long long i;
     unsigned char cipher[16] = {0};
     clock_t program_start, program_end;
     double ts;
@@ -88,7 +88,7 @@ void test_wbsm4_standard_perf() {
     WBCRYPTO_wbsm4_gen_table(wbsm4_ctx, key, sizeof(key));
 
     program_start = clock();
-    for (i = 0; i < TEST_CYCLE_NUM * 64 * 1024; i++) {
+    for (i = 0; i < (long long)TEST_CYCLE_NUM * 64 * 1024; i++) {
         WBCRYPTO_wbsm4_encrypt(msg, cipher, wbsm4_ctx);
     }
     program_end = clock();
